Use alpha address register
The alpha address register now has indexed all of the addresses used in
this demo, so we can switch to it.

This showed a bug in factory.py: since we created the old address
register, we've decided that empty fields actually don't exist.  This
means that `if s['property']` threw a KeyError on addresses with an
empty property field.  This has been changed to `if 'property' in s`
which does not have the same problem.